pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.8"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.119.0",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
]
